+++ sismic/helpers.py
@@ -35,8 +35,7 @@ def run_in_background(interpreter: Interpreter,
                       delay: float=0.05,
                       callback: Callable[[List[MacroStep]], Any]=None) -> threading.Thread:
     """
-    Run given interpreter in background. The time is updated according to
-    *time.time() - starttime*. The interpreter is ran until it reaches a final configuration.
+    Run given interpreter in background. The interpreter is ran until it reaches a final configuration.
     You can manually stop the thread using the added *stop* of the returned Thread object.
     This is for convenience only and should be avoided, because a call to *stop* puts the interpreter in
     an empty (and thus final) configuration, without properly leaving the active states.
@@ -51,7 +50,6 @@ def run_in_background(interpreter: Interpreter,
     def _task():
         starttime = time.time()
         while not interpreter.final:
-            interpreter.time = time.time() - starttime
             steps = interpreter.execute()
             if callback:
                 callback(steps)