+++ sismic/interpreter/__init__.py
@@ -1,4 +1,5 @@
 from .default import Interpreter
+from .clock import Clock
 from ..model.events import Event, InternalEvent, MetaEvent
 
-__all__ = ['Interpreter', 'Event', 'InternalEvent', 'MetaEvent']
+__all__ = ['Interpreter', 'Clock', 'Event', 'InternalEvent', 'MetaEvent']