+++ tests/test_property.py
@@ -20,9 +20,9 @@ class TestInterpreterMetaEvents:
         return property
 
     def test_synchronised_time(self, microwave, property_statechart):
-        assert microwave.time == property_statechart.time
-        microwave.time += 10
-        assert microwave.time == property_statechart.time
+        assert microwave.clock.time == property_statechart.clock.time
+        microwave.clock.time += 10
+        assert microwave.clock.time == property_statechart.clock.time
 
     def test_empty_step(self, microwave, property_statechart):
         microwave.execute()